Mayowa/build premise cameraview artifact (#1)
## Overview
This is a slightly modified version of the Otalias cameraview library. We set a fixed framerate of 15 which is compatible on all devices. This change is in the `FullVideoRecorder` class. Other changes in this PR are to (1) Generate a premise artifact and deploy (2) Remove failing tests (3) Remove code coverage report